pyproject.toml
[tool.black]
line-length = 105
target-version = ['py38']
skip-string-normalization = true

[tool.isort]
line_length = 105
multi_line_output = 3
lines_after_imports = 2
include_trailing_comma = true

[tool.pytest.ini_options]
minversion = '6.0'
addopts = '-s --maxfail=3 --cov=url_cutter'
testpaths = [
    'tests',
]
env = [
    'MAIN_DB_URI=postgresql://postgres:@127.0.0.1:5432/test',
    'CACHE_DB_URI=redis://127.0.0.1:6379/0',
    'STORAGE_URL=http://127.0.0.1/v1/urls/',
    'UID_LEN=2',
    'CELERY_BROKER=redis://127.0.0.1:6379/1',
    'CELERY_BACKEND=redis://127.0.0.1:6379/2',
    'CELERY_ALWAYS_EAGER=True',
    'URL_TTL_SEC=1',
    'PERIOD_CLEARING_EXPIRED_URLS_SEC=2',
    'API_KEY_TTL_SEC=2',
    'PERIOD_CLEARING_EXPIRED_API_KEYS_SEC=2',
    # 'SENTRY_DSN=https://1@2.ingest.sentry.io/3',
]

[tool.poetry]
name = 'url_cutter'
version = '0.1.0'
description = 'url cutter'
authors = ['Sanhoyzer']
readme = 'README.md'
repository = 'https://github.com/Sanchoyzer/url_cutter'

[tool.poetry.dependencies]
python = '~3.8'
fastapi = '0.63.0'
fastapi-utils = '0.2.1'
gunicorn = '20.0.4'
uvicorn = {extras = ['standard'], version = '0.13.4'}
psycopg2-binary = '2.8.6'
redis = '3.5.3'
alembic = '~1.5.6'
SQLAlchemy = '1.3.23'
SQLAlchemy-Utils = '0.36.8'
requests = '~2.25.1'
celery = {extras = ['redis'], version = '5.0.5'}
qrcode = {extras = ['pil'], version = '6.1'}
sentry-sdk = '~1.0.0'

[tool.poetry.dev-dependencies]
black = '^20.8b1'
isort = '^5.7.0'
pytest = '^6.2.2'
pytest-cov = '^2.11.1'
pytest-env = '^0.6.2'
pytest-repeat = '^0.9.1'
pytest-sugar = '^0.9.4'
Faker = '^6.6.0'
molotov = '^2.1'
factory-boy = '^3.2.0'

[build-system]
requires = ['poetry-core>=1.0.0']
build-backend = 'poetry.core.masonry.api'
